Split types out of Builtin.hs
