+++ examples/Main.hs
@@ -8,6 +8,8 @@ import Test.Tasty.HUnit
 import Plutarch (ClosedTerm, POpaque, compile, printScript, printTerm)
 import Plutarch.Bool (PBool (PTrue), pif, (£==))
 import qualified Plutarch.Builtin as B
+import qualified Plutarch.Builtin.List as BL
+import qualified Plutarch.Builtin.Pair as BP
 import Plutarch.ByteString (phexByteStr)
 import Plutarch.Either (PEither (PLeft, PRight))
 import Plutarch.Evaluate (evaluateScript)
@@ -45,26 +47,26 @@ fib = phoistAcyclic $
         1
         $ self £ (n - 1) + self £ (n - 2)
 
-fibs :: Term s (PInteger :--> B.PList PInteger)
+fibs :: Term s (PInteger :--> BL.PList PInteger)
 fibs = phoistAcyclic $
   pfix £$ plam $ \self n ->
     pif
       (n £== 0)
-      (B.singleton £ 0)
+      (BL.singleton £ 0)
       $ pif
         (n £== 1)
-        (B.mkList [1, 0])
+        (BL.mkList [1, 0])
         $ plet (self £ (n - 1)) $ \a ->
           plet (self £ (n - 2)) $ \b ->
-            B.cons
-              (B.headL a + B.headL b)
+            BL.cons
+              (BL.head a + BL.head b)
               a
 
 swap :: (a ~ POpaque, b ~ POpaque) => Term s (B.PPair a b :--> B.PPair b a)
 swap =
   plam $ \p ->
-    B.matchPair p $ \(B.PPair x y) ->
-      B.mkPairData y x
+    BP.matchPair p $ \(B.PPair x y) ->
+      BP.mkPairData y x
 
 uglyDouble :: Term s (PInteger :--> PInteger)
 uglyDouble = plam $ \n -> plet n $ \n1 -> plet n1 $ \n2 -> n2 + n2
@@ -106,10 +108,10 @@ tests =
     , testCase "pfix" $ (printTerm pfix) @?= "(program 1.0.0 ((\\i0 -> i1) (\\i0 -> (\\i0 -> i2 (\\i0 -> i2 i2 i1)) (\\i0 -> i2 (\\i0 -> i2 i2 i1)))))"
     , testCase "fib" $ (printTerm fib) @?= "(program 1.0.0 ((\\i0 -> (\\i0 -> (\\i0 -> i1) (i1 (\\i0 -> \\i0 -> force (i4 (equalsInteger i1 0) (delay 0) (delay (force (i4 (equalsInteger i1 1) (delay 1) (delay (addInteger (i2 (subtractInteger i1 1)) (i2 (subtractInteger i1 2))))))))))) (\\i0 -> (\\i0 -> i2 (\\i0 -> i2 i2 i1)) (\\i0 -> i2 (\\i0 -> i2 i2 i1)))) (force ifThenElse)))"
     , testCase "fib 9 == 34" $ equal (fib £ 9) (34 :: Term s PInteger)
-    , testCase "fibs 3 == [..]" $ equal (fibs £ 5) (B.mkList (reverse [0, 1, 1, 2, 3, 5]) :: Term s (B.PList PInteger))
+    , testCase "fibs 3 == [..]" $ equal (fibs £ 5) (BL.mkList (reverse [0, 1, 1, 2, 3, 5]) :: Term s (B.PList PInteger))
     , testCase "swap (1,2) == (2,1)" $
         let f n = B.IData B.#£ n
-         in equal (swap £ B.mkPairData (f 1) (f 2)) (B.mkPairData (f 2) (f 1) :: Term s (B.PPair POpaque POpaque))
+         in equal (swap £ BP.mkPairData (f 1) (f 2)) (BP.mkPairData (f 2) (f 1) :: Term s (B.PPair POpaque POpaque))
     , testCase "uglyDouble" $ (printTerm uglyDouble) @?= "(program 1.0.0 (\\i0 -> addInteger i1 i1))"
     , testCase "1 + 2 == 3" $ equal (1 + 2 :: Term s PInteger) (3 :: Term s PInteger)
     , testCase "fails: perror" $ fails perror
